+++ CHANGELOG.md
@@ -5,6 +5,19 @@ Changelog
 ------------
 -
 
+[v2.0.20] - 2021-10-28
+-----------------
+[GitHub Release Page](https://github.com/ToranSharma/Duo-Strength/releases/tag/v2.0.20)
+### Fixed
+- Detection of unloaded skills on page. If any are found, they are forced to be
+  loaded by scaling the tree so that it is all on screen. The rest of the
+  features can then work with the fully loaded tree.
+
+### Added
+- Loading animation that covers the visual effects of the forced skill loading,
+  as well as showing that Duo Strength is requesting the data needed to
+  function.
+
 [v2.0.19] - 2021-10-28
 -----------------
 [GitHub Release Page](https://github.com/ToranSharma/Duo-Strength/releases/tag/v2.0.19)
@@ -1393,6 +1406,7 @@ Changelog
   from a lesson to the main page.
 
 [Unreleased]: https://github.com/ToranSharma/Duo-Strength/compare/master...develop
+[v2.0.20]: https://github.com/ToranSharma/Duo-Strength/compare/v2.0.19...v2.0.20
 [v2.0.19]: https://github.com/ToranSharma/Duo-Strength/compare/v2.0.18...v2.0.19
 [v2.0.18]: https://github.com/ToranSharma/Duo-Strength/compare/v2.0.17...v2.0.18
 [v2.0.17]: https://github.com/ToranSharma/Duo-Strength/compare/v2.0.16...v2.0.17